+++ app/src/androidTest/java/org/simple/clinic/rules/ServerAuthenticationRule.kt
@@ -197,7 +197,7 @@ class ServerAuthenticationRule : TestRule {
 
   private fun verifyUserCanSyncData() {
     val (loggedInUser) = userSession.loggedInUser().blockingFirst()
-    assertThat(userSession.isUserLoggedIn()).isTrue()
+    assertThat(userSession.isUserPresentLocally()).isTrue()
     assertThat(loggedInUser!!.status).isEqualTo(UserStatus.ApprovedForSyncing)
     assertThat(loggedInUser.loggedInStatus).isEqualTo(User.LoggedInStatus.LOGGED_IN)
   }